Moved cli logic got own module
